Restructured to split views.py into views.py and views_api.py in cohorts and idc_collections. Added cohorts/utils_api.py. Various changes to return results in a consistent manner across APIs
