+++ cohorts/utils.py
@@ -71,66 +71,6 @@ def _delete_cohort(user, cohort_id):
     return cohort_info
 
 
-def _save_cohort_api(user, name, data, case_insens=True):
-
-    description = data['description']
-    filterset = data['filterSet']
-    attributes = filterset["attributes"]
-    cohort_id = 'cohort_id' in data and data['cohort_id'] or None
-
-
-    if not filterset or not name:
-        # Can't save/edit a cohort when nothing is being changed!
-        return { 'message': "Can't save a cohort with no information to save! (Name and filters not provided.)" }
-
-    blacklist = re.compile(BLACKLIST_RE, re.UNICODE)
-    match = blacklist.search(str(name))
-    if match:
-        # XSS risk, log and fail this cohort save
-        match = blacklist.findall(str(name))
-        logger.error('[ERROR] While saving a cohort, saw a malformed name: ' + name + ', characters: ' + str(match))
-        return {'message': "Your cohort's name contains invalid characters; please choose another name."}
-
-    # If we're only changing the name, just edit the cohort and update it
-    if cohort_id:
-        cohort = Cohort.objects.get(id=cohort_id)
-        cohort.name = name
-        cohort.save()
-        return {'cohort_id': cohort.id}
-
-    # Make and save cohort
-    cohort = Cohort.objects.create(name=name, description=description)
-    cohort.save()
-
-    perm = Cohort_Perms(cohort=cohort, user=user, perm=Cohort_Perms.OWNER)
-    perm.save()
-
-    # For now, any set of filters in a cohort is a single 'group'; this allows us to, in the future,
-    # let a user specify a different operator between groups (eg. (filter a AND filter b) OR (filter c AND filter D)
-    grouping = Filter_Group.objects.create(resulting_cohort=cohort, operator=Filter_Group.AND)
-
-    # Get versions of datasets to be filtered, and link to filter group
-    imaging_version = 'imaging_version' in filterset and \
-                      len(DataVersion.objects.filter(name='TCIA Image Data', version=filterset['imaging_version'])) == 1 and \
-                      DataVersion.objects.get(name='TCIA Image Data', version=filterset['imaging_version']) or \
-                      DataVersion.objects.get(active=True, name='TCIA Image Data')
-    grouping.data_versions.add(imaging_version)
-
-    bioclin_version = 'bioclin_version' in filterset and \
-                      len(DataVersion.objects.filter(name='TCGA Clinical and Biospecimen Data', version=filterset['bioclin_version'])) == 1 and \
-                      DataVersion.objects.get(name='TCGA Clinical and Biospecimen Data', version=filterset['bioclin_version']) or \
-                      DataVersion.objects.get(active=True, name='TCGA Clinical and Biospecimen Data')
-    grouping.data_versions.add(bioclin_version)
-
-
-    for attr in attributes:
-        filter_values = attributes[attr]
-        attr_id = Attribute.objects.get(name=attr)
-        Filters.objects.create(resulting_cohort=cohort, attribute=attr_id, value=",".join(filter_values), filter_group=grouping).save()
-
-    return {'cohort_id': cohort.id}
-
-
 def _save_cohort(user, filters=None, name=None, cohort_id=None, case_insens=True):
 
     if not filters and not name:
