+++ idc_collections/views/__init__.py
@@ -1,0 +1,2 @@
+from .views import *
+from .views_api import *
