+++ idc_collections/views/views.py
@@ -33,7 +33,7 @@ from django.urls import reverse
 from collections import OrderedDict
 from data_upload.models import UserUpload, UserUploadedFile
 from idc_collections.models import User_Feature_Definitions, User_Feature_Counts, \
-    Program, Collection, Attribute, DataVersion
+    Program, Collection
 from solr_helpers import *
 from sharing.service import create_share
 from googleapiclient.errors import HttpError
@@ -70,20 +70,6 @@ def program_list(request, is_public=False, is_API=False):
     return render(request, template, context)
 
 
-def public_program_list_api(request):
-
-    programs = Program.objects.filter(is_public=True)
-
-    programs_info = [{
-                "name": program.name,
-                "short_name": program.short_name,
-                "description": program.description,
-                "active": program.active} for program in programs]
-
-    return JsonResponse(programs_info, safe=False)
-#    return HttpResponse(programs_info,  content_type='application/json')
-
-
 @login_required
 def program_detail(request, program_id=0, is_API=False):
     # """ if debug: logger.debug('Called ' + sys._getframe().f_code.co_name) """
@@ -109,109 +95,6 @@ def program_detail(request, program_id=0, is_API=False):
         'shared': shared
     }
     return render(request, template, context)
-
-
-def program_detail_api(request, program_name=None ):
-    # """ if debug: logger.debug('Called ' + sys._getframe().f_code.co_name) """
-    programs = Program.objects.filter(is_public=True, active=True).distinct()
-    program = programs.get(short_name__iexact=program_name)
-
-    collections = program.collection_set.all()
-
-    collections_info = []
-    for collection in collections:
-        dvs = collection.data_versions.all()
-        data = {
-            "name": collection.name,
-            "short_name": collection.short_name,
-            "description": collection.description,
-            "active": collection.active,
-            "is_public": collection.is_public,
-            "owner_id": collection.owner_id,
-            "data_version": [{"name":dv.name.replace(' ','_'),"data_type":dv.data_type, "version":dv.version} for dv in dvs] }
-        collections_info.append(data)
-
-    return JsonResponse(collections_info, safe=False)
-
-
-def collection_detail_api(request, program_name, collection_name):
-    # """ if debug: logger.debug('Called ' + sys._getframe().f_code.co_name) """
-
-    collection_info = {}
-    try:
-        collection = Collection.objects.get(name=collection_name)
-
-    except ObjectDoesNotExist as e:
-        collection_info = {
-            "message": "Collection {} does not exist".format(collection_name),
-            "code": "",
-            "not_found": ""
-        }
-    else:
-
-        attribute_group = request.GET['attribute_group']
-        version = request.GET['version']
-
-        try:
-            if 'version' in request.GET:
-                dataVersion = collection.data_versions.get(name=attribute_group, version=version)
-            else:
-                dataVersion = collection.data_versions.get(name=attribute_group, active=True)
-
-        except ObjectDoesNotExist as e:
-            collection_info = {
-                "message": "Attribute group/version {}/{} does not exist".format(attribute_group, version),
-                "code": "",
-                "not_found": ""
-            }
-        else:
-            try:
-                # # Get attributes that are not cross collection but collection_name
-                # collection = Program.objects.get(short_name__iexact=program_name).collection_set.get(
-                #     short_name__iexact=collection_name)
-                # data_version = collection.data_versions.get(name=attribute_group, version=version)
-
-                # if data_version.data_type == 'A':
-                #     bq_tables = dataVersion.bigquerytable_set.all()
-                # else:
-                #     bq_tables = dataVersion.bigquerytable_set.filter(name__icontains=collection_name)
-
-
-                # bq_tables = dataVersion.bigquerytable_set.all()
-                bq_tables = dataVersion.datasource_set.filter(source_type='B')
-
-                fields = []
-                for table in bq_tables:
-                    for attribute in table.attribute_set.all():
-                        fields.append({
-                            "id": attribute.id,
-                            "name": attribute.name,
-                            "display_name": attribute.display_name,
-                            "description": attribute.description,
-                            "data_type": attribute.data_type,
-                            "active": attribute.active,
-                            "preformatted_values": attribute.preformatted_values,
-                            "bq_table": table.name
-                            # } for attribute in attributes ]
-                        })
-
-                collection_info = {
-                    "collection_name": collection_name,
-                    "attribute_group": attribute_group,
-                    "version": version,
-                    "active": DataVersion.objects.get(name=attribute_group, version=version).active,
-                    "fields": fields
-                }
-
-            except ObjectDoesNotExist as e:
-                collection_info = {
-                    "message": "Program/collection {}/{} does not exist".format(program_name, collection_name),
-                    "code": "",
-                    "not_found": ""
-                }
-
-    # return HttpResponse(collection_info, content_type='application/json')
-    return JsonResponse(collection_info)
 
 
 def program_upload_existing(request):
